requirements.txt
# Versi 1.1 - Memaksa reinstall
streamlit
faiss-cpu
sentence-transformers
gtts
nltk
requests
moviepy==1.0.3
Pillow
python-dotenv
numpy
deep-translator